requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
boto3==1.4.4
botocore==1.5.38
bs4==0.0.1
click==6.7
dj-database-url==0.4.2
Django==1.10.6
django-cas-ng==3.5.6
django-filter==0.15.2
djangorestframework==3.4.7
docutils==0.13.1
futures==3.1.1
gunicorn==19.7.1
jmespath==0.9.2
Markdown==2.6.7
packaging==16.8
pbr==1.8.1
Pillow==3.2.0
psycopg2==2.7.1
pyparsing==2.2.0
python-cas==1.2.0
python-dateutil==2.6.0
python-dotenv==0.6.4
python-http-client==2.2.1
pytz==2017.2
requests==2.13.0
s3transfer==0.1.10
sendgrid==3.6.5
sendgrid-django==4.0.4
six==1.10.0
stevedore==1.12.0
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
whitenoise==3.3.0
